Renamed StudentAI to Greedy so we can make copy and name it MinMax
