+++ uvschess/uvschess/StudentAIGreedy/packages.config
