+++ uvschess/uvschess/UvsChess.sln
@@ -1,12 +1,12 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25123.0
+# Visual Studio 15
+VisualStudioVersion = 15.0.28307.329
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Framework", "Framework\Framework.csproj", "{BE8F4EFB-458E-431A-A763-5988804DC0ED}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExampleAI", "ExampleAI\ExampleAI.csproj", "{5B942207-A326-4144-850C-CA867E22721A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StudentAI", "StudentAI\StudentAI.csproj", "{5B942208-A326-4144-850C-CA867E22721A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StudentAIGreedy", "StudentAI\StudentAIGreedy.csproj", "{5B942208-A326-4144-850C-CA867E22721A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -29,5 +29,8 @@ Global
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {C600F0B4-C8B3-4765-9591-62E49A941E03}
 	EndGlobalSection
 EndGlobal
